+++ CHANGELOG.md
@@ -5,8 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to
 [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] - 2018-03-21
+ * Allow the `IDocumentStore` to accept an `IUploadConfig` file to pass additional configuration properties
+   while uploading a file.
+
 ## [1.2.0] - 2017-10-06
- * Add a new `S3DocumentStore.getUrl(final String keyName)` function to create absolute Amazon S3 urls ; 
+ * Add a new `S3DocumentStore.getUrl(final String keyName)` function to create absolute Amazon S3 urls ;
  * Add an implementation for `S3Mock.deleteObject(final DeleteObjectRequest deleteObjectRequest)`.
 
 ## [1.1.0] - 2017-09-29